WIP: bring back dialogs
